Fix the downloader
With all databases enabled the ThreadPool queue is exceed (25>16).
This normally isn't an issue for other users of the pool as it will get pushed onto the main thread.
But Android prohibits network activity on the main thread, which the downloader runs from.

Increase it 32 which should be suitable for now.
Maybe consider adding a separate pool in future.

Fixes https://github.com/Divested-Mobile/Hypatia/issues/34
Fixes https://gitlab.com/divested-mobile/hypatia/-/issues/31

Signed-off-by: Tad <tad@spotco.us>